refactor(gym): isolate dynamic settling functor
